feat(timeoutWith): add higher-order lettable version of timeoutWith
